Fix CORS problem
In order to not violate CORS regulation on web browser,`Access-Control-Allow-Origin: *` header needs to be set while responding to client's `POST` request.
And before `POST` request, there's an another `OPTION` request called preflight request [1], the same header above also needs to be set while responding to this kind of request.

[1] https://developer.mozilla.org/en-US/docs/Glossary/Preflight_request