Fix response type error
When the client is requesting the root path of the server, it won't
get the response message, due to the internal server error, it is
caused by the response with wrong data type, the response data
only accept the "string".

To fix this problem, convert the data to the "string" type.